+++ test/base/src/androidTest/kotlin/io/realm/kotlin/test/shared/InitialDataTests.kt
@@ -31,9 +31,11 @@ import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
+import kotlin.test.fail
 
 /**
  * Class testing [io.realm.Configuration.initialDataCallback] functionality.
@@ -69,8 +71,24 @@ class InitialDataTests {
         }.build()
 
         assertFailsWithMessage<RuntimeException>("Boom!") {
+            Realm.open(config).use {
+                fail()
+            }
+        }
+    }
+
+    @Test
+    fun initialData_failureDeletesRealm() {
+        @Suppress("TooGenericExceptionThrown")
+        val config = configBuilder.initialData {
+            assertTrue(fileExists(this.configuration.path))
+            throw RuntimeException("Boom!")
+        }.build()
+        assertFalse(fileExists(config.path))
+        assertFailsWithMessage<RuntimeException>("Boom!") {
             Realm.open(config)
         }
+        assertFalse(fileExists(config.path))
     }
 
     @Test